+++ tests/test_basic.py
@@ -10,6 +10,7 @@
 """
 
 import re
+import sys
 import time
 import uuid
 from datetime import datetime
@@ -1289,6 +1290,14 @@ def test_jsonify_mimetype(app, req_ctx):
     assert rv.mimetype == "application/vnd.api+json"
 
 
+@pytest.mark.skipif(sys.version_info < (3, 7), reason="requires Python >= 3.7")
+def test_json_dump_dataclass(app, req_ctx):
+    from dataclasses import make_dataclass
+    Data = make_dataclass("Data", [("name", str)])
+    value = flask.json.dumps(Data("Flask"), app=app)
+    value = flask.json.loads(value, app=app)
+    assert value == {"name": "Flask"}
+
 def test_jsonify_args_and_kwargs_check(app, req_ctx):
     with pytest.raises(TypeError) as e:
         flask.jsonify("fake args", kwargs="fake")
@@ -1390,6 +1399,17 @@ def test_static_files(app, client):
 
 def test_static_url_path():
     app = flask.Flask(__name__, static_url_path="/foo")
+    app.testing = True
+    rv = app.test_client().get("/foo/index.html")
+    assert rv.status_code == 200
+    rv.close()
+
+    with app.test_request_context():
+        assert flask.url_for("static", filename="index.html") == "/foo/index.html"
+
+
+def test_static_url_path_with_ending_slash():
+    app = flask.Flask(__name__, static_url_path="/foo/")
     app.testing = True
     rv = app.test_client().get("/foo/index.html")
     assert rv.status_code == 200
@@ -1896,21 +1916,24 @@ def test_run_server_port(monkeypatch, app):
 
 
 @pytest.mark.parametrize(
-    "host,port,expect_host,expect_port",
+    "host,port,server_name,expect_host,expect_port",
     (
-        (None, None, "pocoo.org", 8080),
-        ("localhost", None, "localhost", 8080),
-        (None, 80, "pocoo.org", 80),
-        ("localhost", 80, "localhost", 80),
+        (None, None, "pocoo.org:8080", "pocoo.org", 8080),
+        ("localhost", None, "pocoo.org:8080", "localhost", 8080),
+        (None, 80, "pocoo.org:8080", "pocoo.org", 80),
+        ("localhost", 80, "pocoo.org:8080", "localhost", 80),
+        ("localhost", 0, "localhost:8080", "localhost", 0),
+        (None, None, "localhost:8080", "localhost", 8080),
+        (None, None, "localhost:0", "localhost", 0),
     ),
 )
-def test_run_from_config(monkeypatch, host, port, expect_host, expect_port, app):
+def test_run_from_config(monkeypatch, host, port, server_name, expect_host, expect_port, app):
     def run_simple_mock(hostname, port, *args, **kwargs):
         assert hostname == expect_host
         assert port == expect_port
 
     monkeypatch.setattr(werkzeug.serving, "run_simple", run_simple_mock)
-    app.config["SERVER_NAME"] = "pocoo.org:8080"
+    app.config["SERVER_NAME"] = server_name
     app.run(host, port)